search tab header move to json.  backwards compat for binary serializer.
