+++ AndriodApp1/TabsPagerAdapter.cs
@@ -381,163 +381,10 @@ namespace AndriodApp1
         }
     }
 
-    [Serializable]
-    public class SavedStateSearchTabHeader
-    {
-        string LastSearchTerm;
-        long LastRanTime;
-        int LastSearchResultsCount;
 
-        /// <summary>
-        /// Get what you need to display the tab (i.e. result count, term, last ran)
-        /// </summary>
-        /// <param name="searchTab"></param>
-        /// <returns></returns>
-        public static SavedStateSearchTabHeader GetSavedStateHeaderFromTab(SearchTab searchTab)
-        {
-            SavedStateSearchTabHeader searchTabState = new SavedStateSearchTabHeader();
-            searchTabState.LastSearchResultsCount = searchTab.LastSearchResultsCount;
-            searchTabState.LastSearchTerm = searchTab.LastSearchTerm;
-            searchTabState.LastRanTime = searchTab.LastRanTime.Ticks;
-            return searchTabState;
-        }
 
-        /// <summary>
-        /// these by definition will always be wishlist tabs...
-        /// this restores the wishlist tabs, optionally with the search results, otherwise they will be added later.
-        /// </summary>
-        /// <param name="savedState"></param>
-        /// <returns></returns>
-        public static SearchTab GetTabFromSavedState(SavedStateSearchTabHeader savedStateHeader, List<SearchResponse> responses)
-        {
-            SearchTab searchTab = new SearchTab();
-            searchTab.SearchResponses = responses;
-            searchTab.LastSearchTerm = savedStateHeader.LastSearchTerm;
-            searchTab.LastRanTime = new DateTime(savedStateHeader.LastRanTime);
-            searchTab.SearchTarget = SearchTarget.Wishlist;
-            searchTab.LastSearchResultsCount = responses != null ? responses.Count : savedStateHeader.LastSearchResultsCount;
-            if (SearchFragment.FilterSticky)
-            {
-                searchTab.FilterSticky = SearchFragment.FilterSticky;
-                searchTab.FilterString = SearchFragment.FilterStickyString;
-                SearchFragment.ParseFilterString(searchTab);
-            }
-            searchTab.SortHelper = new SortedDictionary<SearchResponse, object>(new SearchResultComparableWishlist(searchTab.SortHelperSorting));
-            if (responses != null)
-            {
-                foreach (SearchResponse resp in searchTab.SearchResponses)
-                {
-                    if (!searchTab.SortHelper.ContainsKey(resp))
-                    {
-                        //bool isItActuallyNotThere = true;
-                        //foreach(var key in searchTab.SortHelper.Keys)
-                        //{
-                        //    if (key.Username == resp.Username)
-                        //    {
-                        //        if ((key.FileCount == resp.FileCount) && (key.LockedFileCount == resp.LockedFileCount))
-                        //        {
-                        //            if (key.FileCount != 0 && (key.Files.First().Filename == resp.Files.First().Filename))
-                        //            {
-                        //                isItActuallyNotThere = false;
-                        //            }
-                        //            if (key.LockedFileCount != 0 && (key.LockedFiles.First().Filename == resp.LockedFiles.First().Filename))
-                        //            {
-                        //                isItActuallyNotThere = false;
-                        //            }
-                        //        }
-                        //    }
-                        //}
 
-                        searchTab.SortHelper.Add(resp, null);
-                    }
-                    else
-                    {
-
-                    }
-                }
-            }
-            return searchTab;
-
-        }
-    }
-
-    [Serializable]
-    public class SavedStateSearchTab
-    {
-        //there are all of the things we must save in order to later restore a SearchTab
-        public List<SearchResponse> searchResponses;
-        public string LastSearchTerm;
-        public long LastRanTime;
-        public static SavedStateSearchTab GetSavedStateFromTab(SearchTab searchTab)
-        {
-            SavedStateSearchTab searchTabState = new SavedStateSearchTab();
-            searchTabState.searchResponses = searchTab.SearchResponses.ToList();
-            searchTabState.LastSearchTerm = searchTab.LastSearchTerm;
-            searchTabState.LastRanTime = searchTab.LastRanTime.Ticks;
-            return searchTabState;
-        }
-
-        /// <summary>
-        /// these by definition will always be wishlist tabs...
-        /// </summary>
-        /// <param name="savedState"></param>
-        /// <returns></returns>
-        public static SearchTab GetTabFromSavedState(SavedStateSearchTab savedState, bool searchResponsesOnly = false, SearchTab oldTab = null)
-        {
-            SearchTab searchTab = new SearchTab();
-            searchTab.SearchResponses = savedState.searchResponses;
-            if (!searchResponsesOnly)
-            {
-                searchTab.LastSearchTerm = savedState.LastSearchTerm;
-                searchTab.LastRanTime = new DateTime(savedState.LastRanTime);
-            }
-            else
-            {
-                searchTab.LastSearchTerm = oldTab.LastSearchTerm;
-                searchTab.LastRanTime = oldTab.LastRanTime;
-            }
-            searchTab.SearchTarget = SearchTarget.Wishlist;
-            searchTab.LastSearchResultsCount = searchTab.SearchResponses.Count;
-            if (SearchFragment.FilterSticky)
-            {
-                searchTab.FilterSticky = SearchFragment.FilterSticky;
-                searchTab.FilterString = SearchFragment.FilterStickyString;
-                SearchFragment.ParseFilterString(searchTab);
-            }
-            searchTab.SortHelper = new SortedDictionary<SearchResponse, object>(new SearchResultComparableWishlist(searchTab.SortHelperSorting));
-            foreach (SearchResponse resp in searchTab.SearchResponses)
-            {
-                if (!searchTab.SortHelper.ContainsKey(resp))
-                {
-                    //bool isItActuallyNotThere = true;
-                    //foreach(var key in searchTab.SortHelper.Keys)
-                    //{
-                    //    if (key.Username == resp.Username)
-                    //    {
-                    //        if ((key.FileCount == resp.FileCount) && (key.LockedFileCount == resp.LockedFileCount))
-                    //        {
-                    //            if (key.FileCount != 0 && (key.Files.First().Filename == resp.Files.First().Filename))
-                    //            {
-                    //                isItActuallyNotThere = false;
-                    //            }
-                    //            if (key.LockedFileCount != 0 && (key.LockedFiles.First().Filename == resp.LockedFiles.First().Filename))
-                    //            {
-                    //                isItActuallyNotThere = false;
-                    //            }
-                    //        }
-                    //    }
-                    //}
-
-                    searchTab.SortHelper.Add(resp, null);
-                }
-                else
-                {
-
-                }
-            }
-            return searchTab;
-        }
-    }
+ 
 
 
     public class SearchTabItemRecyclerAdapter : RecyclerView.Adapter
